requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
PyMuPDF==1.23.8
python-docx==0.8.11
google-generativeai==0.3.0
groq==0.4.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pypdf2==3.0.1